demrepofdave/al5int/fix_bugs_and_improve_code_quality: Brokeout menu code into multiple files
